updated files related to SwapStrategy.sol to load by id

+++ src/market-making/leaves/Fee.sol
@@ -45,8 +45,8 @@ library Fee {
     )
         internal
         pure
-        returns (uint256 amount)
+        returns (uint128 amount)
     {   
-        amount = mulDiv(totalAmount, portion, denominator);
+        amount = uint128(mulDiv(totalAmount, portion, denominator));
     }
 }
